+++ resources/views/Akunsiswa/akunsiswa.blade.php
@@ -64,7 +64,7 @@
                     <td>{{ $item->level}}</td>
                     <td>
                         <a href="edit-akunsiswa/{{$item->id}}" class="btn btn-info">Edit</a>
-                        <a href="delete-akunsiswa/{{$item->id}}" class="btn btn-danger">Delete</a>
+                        <a href="delete-akunsiswa/{{$item->id}}" onclick="return confirm('Apakah Anda Yakin Menghapus Akun Siswa?');"  class="btn btn-danger">Delete</a>
                     </td>
                 </tr>
                 @endforeach
